* release/0.7.0: (97 commits)
  Fix Grid Partitioner selection
  Fix warning
  Cosmetics
  Remove ObjectRegistry as it is not thread safe. The eckit class ClassExtent can be used instead
  move atlas::grid::deprecated::RotatedGrid to mir::util::
  Fix usage of Trans partitioner
  Fix warnings
  Fix possible out-of-bounds
  Fix typo
  Missing Fortran API for gather/scatter Spectral fieldset
  Documentation and cosmetics
  Bump version to 0.7.0
  Compatibility with eckit [0.12.3]
  ATLAS-82 Completely removed Resource from atlas library (still in some apps awaiting ECKIT-145)
  Compatibility with eckit 0.13.0
  ATLAS-82 Remove reading of resource files
  ATLAS-82 Remove Resource from Gmsh and mesh::generators::Structured
  Increase version string length in Fortran
  Cosmetic changes
  Fix autogenerated diff
  ...

+++ src/atlas/numerics/fvm/Method.cc
@@ -102,7 +102,7 @@ void Method::setup()
       }
       array::ArrayView<double,2> node2edge_sign( nodes_.field("node2edge_sign") );
 
-      atlas_omp_parallel_for( int jnode=0; jnode<nnodes; ++jnode )
+      atlas_omp_parallel_for( size_t jnode=0; jnode<nnodes; ++jnode )
       {
         for(size_t jedge = 0; jedge < node_edge_connectivity.cols(jnode); ++jedge)
         {
